+++ tools/pylib/zoidberg/test_rzline.py
@@ -1,10 +1,10 @@
-from . import poloidal_grid as grid
+from . import rzline
 
 import numpy as np
 
 def test_circular_boundaries():
-    inner = grid.circle(R0=1.0, r=1.0, n=20)
-    outer = grid.circle(R0=1.0, r=2.0, n=20)
+    inner = rzline.circle(R0=1.0, r=1.0, n=20)
+    outer = rzline.circle(R0=1.0, r=2.0, n=20)
     
     assert np.amax(inner.R) <= 2.0
     assert np.amin(inner.R) >= 0.0
@@ -14,8 +14,8 @@ def test_distance():
     # Check the RZline.distance() function
     
     n = 50
-    inner = grid.circle(R0=1.0, r=1.0, n=n)
-    outer = grid.circle(R0=1.0, r=2.0, n=n)
+    inner = rzline.circle(R0=1.0, r=1.0, n=n)
+    outer = rzline.circle(R0=1.0, r=2.0, n=n)
     
     dist = inner.distance()
 
@@ -34,20 +34,11 @@ def test_distance():
     
 def test_order_by_distance():
     # Check the RZline.orderByDistance function
-    inner = grid.circle(R0=1.0, r=1.0, n=20)
-    outer = grid.circle(R0=1.0, r=2.0, n=20)
+    inner = rzline.circle(R0=1.0, r=1.0, n=20)
+    outer = rzline.circle(R0=1.0, r=2.0, n=20)
     
     new_inner = inner.orderByDistance(n=10)
     
     assert len(new_inner.R) == 10
-    
-    
-def test_grid_annulus():
-    n = 50
-    inner = grid.circle(R0=1.0, r=1.0, n=n)
-    outer = grid.circle(R0=1.0, r=2.0, n=n)
-    
-    
-    grid.grid_annulus(inner, outer, 5, 10)
